requirements.txt
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
python-dotenv
sqlalchemy
psycopg2-binary
pandas
